+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 4.0.0a1 (unreleased)
 --------------------
 
+- Register our ``AnnotationStorage`` as ``IImageScaleStorage`` factory.
+  We require ``plone.namedfile >= 6.0.0a2`` for this.
+  Fixes `issue 49 <https://github.com/plone/plone.app.tiles/issues/48>.  [maurits]
+
 - Drop support for Python 2 and Plone 5.1.
   Plone 5.2 and 6.0 are supported, on Python 3.
   See `issue 49 <https://github.com/plone/plone.app.tiles/issues/49>.  [maurits]